pyproject.toml
[project]
name = "divergence-engine"
version = "0.1.0"
description = "Prediction Market × Financial Market Divergence Engine"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "yfinance>=0.2.40",
    "typer>=0.12",
    "rich>=13.0",
    "numpy>=1.26",
    "pandas>=2.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dashboard = [
    "streamlit>=1.35",
    "plotly>=5.22",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[project.scripts]
divergence-engine = "divergence_engine.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["divergence_engine*"]

[tool.ruff]
target-version = "py310"
line-length = 100
